Players 객체 작성 배열관리를 따로 하도록 변경함.
