+++ CardGameApp/CardGameApp/model/CardGame.swift
@@ -12,41 +12,24 @@ class GameTable {
     private let playerEntry: PlayerEntry
     private let studType: StudType
     private var dealer: Dealer
-    private(set) var players: [Player]
+    private(set) var participants: Players
   
     init(playerEntry: PlayerEntry = .two, studType: StudType = .sevenStud) {
         self.playerEntry = playerEntry
         self.studType = studType
         self.dealer = Dealer(name: "Dealer")
-        self.players = [Player]()
-        
-        participatePlayer()
-        initDrawCard()
+        self.participants = Players()
+        setUpGame()
     }
     
-    // 플레이어 세팅
-    private func participatePlayer() {
-        var index = 1
-        playerEntry.each {
-            players.append(Player(name: "Player\(index)"))
-            index += 1
-        }
-        players.append(dealer)
-    }
-    
-    // 플레이어 카드 분배.
-    private func initDrawCard() {
-        players.forEach { (player) in
-            studType.each {
-                let card = dealer.drawCard()
-                player.bringCard(card: card)
-            }
-        }
+    private func setUpGame() {
+        participants.joinPlayerToGame(playerEntry: playerEntry, dealer: dealer)
+        participants.initDrawCard(studType: studType, dealer: dealer)
     }
     
     // 카드개수비교해서 게임 속행할지 확인
     func isContinue() -> Bool {
-        return dealer.cardDeckCount() > players.count
+        return dealer.cardDeckCount() > participants.players.count
     }
 }
